♻️ Extract shared schemas and add tool calls extraction utility
- Move message parsing schemas to shared schema.ts file
- Create extractToolCallsFromMessage utility for parsing tool calls
- Update extractReasoningFromMessage to use shared schema
- Define ToolCalls type for better type safety
